Feat: "리프레시 토큰 관련 슬라이딩 세션 서비스 로직 추가 구현"
현재 구현된 로그인의 경우, 로그인을 시도할 때마다 새로운 리프레시 토큰을 반환하도록 구현되어 있어서,
만약 컴퓨터에서 로그인 하고, 모바일에서 같은 계정에 로그인 한다면, 컴퓨터에 있는 리프레시 토큰이 무효화 되게 된다.
따라서, 로그인 시도할 때, Redis 에 저장된 리프레시 토큰이 있다면, 그 토큰을 반환하고, 또 유효 기간을 연장 시키는 슬라이딩 세션 방식을 섞기로 했다.

이렇게 하면 장점으로 다음과 같은 점이 있다.
* 다른 기기에서 로그인 해도 로그인이 풀리지 않음
* 로그인 및 액세스 토큰 재발급시 리프레시 토큰이 연장되기 때문에 사용자가 30일 이내에 재방문 한다면 다시 로그인 할 필요가 없어져 편리함
* 리프레시 토큰이 유출되어도 Redis에 있는 리프레시 토큰을 지우면, "모든 기기에서 로그아웃" 기능처럼 모든 기기의 리프레시 토큰을 한 번에 만료시킬 수 있음

하지만 단점으로는 다음과 같은 점이 있다.
* 원래 사용자가 토큰 유출을 눈치 채지 못해 로그아웃을 하지 않는다면, 악성 유저가 반영구적으로 해당 계정에 접근할 수 있게 됨

+++ src/main/java/com/filmdoms/community/account/service/AccountService.java
@@ -48,9 +48,12 @@ public class AccountService {
             throw new ApplicationException(ErrorCode.INVALID_PASSWORD);
         }
 
-        log.info("리프레시 토큰 생성");
+        log.info("저장된 토큰 존재 여부 확인, 없다면 생성");
         String key = UUID.nameUUIDFromBytes(email.getBytes()).toString();
-        String refreshToken = jwtTokenProvider.createRefreshToken(key);
+        String refreshToken = refreshTokenRepository.findByKey(key)
+                .orElseGet(() -> jwtTokenProvider.createRefreshToken(key));
+
+        log.info("리프레시 토큰 저장 / 갱신");
         refreshTokenRepository.save(key, refreshToken);
 
         return LoginResponseDto.builder()
@@ -73,6 +76,9 @@ public class AccountService {
             throw new ApplicationException(ErrorCode.INVALID_TOKEN);
         }
 
+        log.info("리프레시 토큰 갱신");
+        refreshTokenRepository.save(key, refreshToken);
+
         log.info("새로운 엑세스 토큰 발급");
         String accessToken = jwtTokenProvider.createAccessToken(key);
         return RefreshAccessTokenResponseDto.builder()
@@ -82,8 +88,17 @@ public class AccountService {
 
     public void logout(String refreshToken) {
 
-        log.info("토큰 내 Subject 추출");
+        log.info("토큰 내 저장된 키 추출");
         String key = jwtTokenProvider.getSubject(refreshToken);
+
+        log.info("키로 저장된 토큰 호출");
+        String savedToken = refreshTokenRepository.findByKey(key)
+                .orElseThrow(() -> new ApplicationException(ErrorCode.TOKEN_NOT_IN_DB));
+
+        log.info("저장된 토큰과 대조");
+        if (!Objects.equals(savedToken, refreshToken)) {
+            throw new ApplicationException(ErrorCode.INVALID_TOKEN);
+        }
 
         log.info("저장된 토큰 삭제");
         refreshTokenRepository.deleteByKey(key);
